requirements.txt
# Django Framework
Django>=4.2.0
djangorestframework>=3.14.0

# Database
psycopg2-binary>=2.9.0
django-environ>=0.10.0

# SSH и сетевые библиотеки
paramiko>=3.3.1
scp>=0.14.5
asyncssh>=2.14.2

# Веб-интерфейс и формы
django-crispy-forms>=2.0.0
crispy-bootstrap5>=0.7.0
django-widget-tweaks>=1.4.12

# Задачи в фоне
celery>=5.3.0
redis>=4.6.0
django-celery-beat>=2.5.0

# WebSocket поддержка
channels>=4.0.0
channels-redis>=4.1.0

# Мониторинг и логирование
django-extensions>=3.2.0
colorlog>=6.7.0

# Безопасность
cryptography>=41.0.7
python-dotenv>=1.0.0
python-dateutil>=2.8.2

# Утилиты
netaddr>=0.9.0
psutil>=5.9.6
tabulate>=0.9.0

# Разработка
django-debug-toolbar>=4.1.0
